feat: added some error handling and made the database class more fool-proof
